requirements.txt
flet[all]=0.28.3
python-dotenv
pandas